+++ pycrystallography/core/__init__.py
@@ -1,23 +1,12 @@
-# coding: utf-8
-# Copyright (c) Pycrystallography Development Team.
-# Distributed under the terms of the MIT License.
+"""Core domain models."""
+from __future__ import annotations
 
-from __future__ import unicode_literals
-#from .orientation import Orientation
-#from .crystalOrientation import CrystalOrientation
-#from .quaternion import Quaternion
-#from .millerDirection import MillerDirection
-#from .millerPlane import MillerPlane
-#from .saedAnalyzer import SaedAnalyzer
-#from .orientationRelation import OrientationRelation
-#from .crystallographyFigure import CrystallographyFigure
+from .models import CompositePattern, OrientationRelation, Phase, Variant, ensure_variants_unique
 
-
-"""
-This package contains core modules and classes for representing orientation and
-operations on them.
-"""
-
-__author__ = "K V Mani Krishna"
-__date__ = "Jul 14, 2017"
-
+__all__ = [
+    "CompositePattern",
+    "OrientationRelation",
+    "Phase",
+    "Variant",
+    "ensure_variants_unique",
+]